% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnivora}
\alias{carnivora}
\title{255 genes trees for 53 carnivora species}
\format{
A list of named trees of class multiPhylo
}
\source{
\url{https://doi.org/10.7554/eLife.63167}
}
\usage{
carnivora
}
\description{
This dataset contains 255 carnivora gene trees for 53 carnivora species. 
It is a subset of the dataset used in Allio et al. (2021), obtained 
by only keeping genes for which all 53 species were present.
}
\examples{
data(carnivora)
class(carnivora)
}
\references{
Allio, R., Tilak, M. K., Scornavacca, C., Avenant, N. L., Kitchener, A. C., Corre, E., ... & Delsuc, F. (2021). High-quality carnivoran genomes from roadkill samples enable comparative species delineation in aardwolf and bat-eared fox. Elife, 10, e63167.
}
\keyword{datasets}
