% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{detect.cell.outliers}
\alias{detect.cell.outliers}
\title{detect.cell.outliers}
\usage{
detect.cell.outliers(mat2WR, k = 3)
}
\arguments{
\item{mat2WR}{the 2WR matrix obtained with the Dist2WR function.}

\item{k}{the strength of outlier assignement. the Higher this value the more stringent the detection (less outliers detected).}
}
\value{
"outcell" All cell-by-cell outliers as a matrix with two columns. Each line represents a cell-by-cell outliers
}
\description{
Function to detect cell outliers (species and genes)
}
\details{
This function must be used after all complete outliers (species and genes) have been removed from the data.
}
