% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{detect.complete.outliers}
\alias{detect.complete.outliers}
\title{detect.complete.outliers}
\usage{
detect.complete.outliers(mat2WR, k = 1.5, thres = 0.5)
}
\arguments{
\item{mat2WR}{the 2WR matrix obtained with the Dist2WR function.}

\item{k}{the strength of outlier assignement. the Higher this value the more stringent the detection (less outliers detected).}

\item{thres}{threshold above which genes or species are considered as complete outliers. 0.5 means that a gene or a species is a complete outlier if it is detected as outlier for more than 50\% of the species or genes respectively.}
}
\value{
"mat2WR" The 2WR matrix used to detect outliers.
"outgn" Array containing all the complete outlier genes detected.
"outsp" Array containing all the complete outlier species detected.
}
\description{
Function to detect complete outliers (species and genes)
}
\details{
Must be runed before the detection of cell outliers
detect.complete.outliers is a function taken from the method phylo-MCOA (de Vienne M.D., Ollier S. et Aguileta G. (2012) Phylo-MCOA: A Fast and Efficient Method to Detect Outlier Genes and Species in Phylogenomics Using Multiple Co-inertia Analysis. Molecular Biology and Evolution 29 : 1587 – 1598).
}
