% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{Dist2WR}
\alias{Dist2WR}
\title{Dist2WR}
\usage{
Dist2WR(Distatis)
}
\arguments{
\item{Distatis}{is the output of the fonction mat2Dist (or of distatis from the Distatis R package (Beaton D., Chin Fatt C., Abdi H. (2013) DistatisR: DISTATIS Three Way Metric Multidimensional Scaling. Package R))}
}
\value{
2WR matrix is a gene x specie matrix. Each cell corresponds to the distance of a specie from a gene tree to the reference position of this specie for every gene trees.
}
\description{
This function creates the two-way reference matrix (2WR) from distatis results.
}
