% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fungi}
\alias{fungi}
\title{246 genes trees for 21 fungal species}
\format{A list of trees of class multiPhylo}
\source{
\url{https://academic.oup.com/sysbio/article/57/4/613/1632046}
}
\usage{
fungi
}
\description{
A dataset containing all 246 gene trees explored in Aguileta et al. 2008, 
for 21 fungal species covering the diversity of fungi.
}
\keyword{datasets}
