% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impMean.R
\name{impMean}
\alias{impMean}
\title{impMean}
\usage{
impMean(matrices)
}
\arguments{
\item{matrices}{A list of distance matrices.}
}
\value{
Returns a list of matrices with same dimensions, with rows and columns 
in the same order and missing data (if any) imputed.
}
\description{
Impute missing data in a list of matrices. Matrices are first given the same dimension, 
then missing entries are computed by taking the average value in non-missing corresponding
entries in all matrices.
}
