% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{impPCA.multi}
\alias{impPCA.multi}
\title{impPCA.multi}
\usage{
impPCA.multi(matrices, ncp = 3, center = FALSE, scale = FALSE,
  maxiter = 10000)
}
\arguments{
\item{matrices}{A list of distance matrices with missing data. Each matrices should be named (use the rename.genes function if it is not the case)}
\item{ncp}{integer corresponding to the number of components used to to predict the missing entries.}
\item{center}{boolean. By default FALSE leading to data not centered.}
\item{scale}{boolean. By default FALSE leading to not a same weight for each variable.}
\item{maxiter}{integer, maximum number of iteration for the algorithm.}
}
\value{
Return a list of matrices without missing data.
}
\description{
Imputing missing data in matrices
}
\seealso{
\code{\link{imputePCA2}}
}
