% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{imputePCA2}
\alias{imputePCA2}
\title{imputePCA2}
\usage{
imputePCA2(X, ncp = 2, center = FALSE, scale = FALSE,
  method = c("Regularized", "EM"), row.w = NULL, coeff.ridge = 1,
  threshold = 1e-06, seed = NULL, nb.init = 1, maxiter = 1000)
}
\arguments{
\item{X}{a data.frame with continuous variables containing missing values}

\item{ncp}{integer corresponding to the number of components used to to predict the missing entries}

\item{center}{boolean. By default FALSE leading to data not centered}

\item{scale}{boolean. By default FALSE leading to not a same weight for each variable}

\item{method}{"Regularized" by default or "EM"}

\item{row.w}{row weights (by default, a vector of 1 for uniform row weights)}

\item{coeff.ridge}{1 by default to perform the regularized imputePCA2 algorithm; useful only if method="Regularized". Other regularization terms can be implemented by setting the value to less than 1 in order to regularized less (to get closer to the results of the EM method) or more than 1 to regularized more (to get closer to the results of the mean imputation)}

\item{threshold}{the threshold for assessing convergence}

\item{seed}{integer, by default seed = NULL implies that missing values are initially imputed by the mean of each variable. Other values leads to a random initialization}

\item{nb.init}{integer corresponding to the number of random initializations; the first initialization is the initialization with the mean imputation}

\item{maxiter}{integer, maximum number of iteration for the algorithm}
}
\description{
Impute the missing values of a dataset with the Principal Components Analysis model.
}
\details{
imputePCA function from missMDA package (Josse J. et Husson F. (2012) Handling missing values in exploratory multivariate data analysis method. Journal de la Société Française de Statistique vol. 153 (2): 79-99.) with some ajustements to fit trees data.
}
\seealso{
\code{\link[missMDA]{imputePCA}}
}
