% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{PhylteR}
\alias{PhylteR}
\title{PhylteR}
\usage{
PhylteR(trees, distance = "patristic", k = 1.5, thres = 0.5,
  gene.names = NULL, Norm = "NONE")
}
\arguments{
\item{trees}{The list of gene trees}

\item{distance}{parameter from the function trees2matrices to transform trees into distance matrices. Distance could be "nodal" or "patristic" (default).}

\item{k}{the strength of outlier assignement. The higher this value the more stringent the detection (less outliers detected).}

\item{thres}{For the detection of complete outlier. Threshold above which genes or species are considered as complete outliers. 0.5 means that a gene or a species is a complete outlier if it is detected as outlier for more than 50\% of the species or genes respectively.}

\item{gene.names}{List of gene names if the user want to renames the list of trees. NULL by default.}

\item{Norm}{Type of normalization used for the function mat2dist. Current options are NONE (default) or MFA (that normalizes each matrix so that its first eigenvalue is equal to one).}
}
\value{
$complete$outgn is the list of complete outlier genes, $complete$outsp is the list of complete outliers species, $CellByCell$outcell is the list of cell outliers.
}
\description{
This function finds complete and cell outliers inside a list of gene trees.
}
\details{
The detection is done in two steps. The first step is the detection on complete outlier. Complete outliers detected are then removed of the trees list and the second step is the detection of cell outliers.
}
