% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{plot2WR}
\alias{plot2WR}
\title{plot2WR}
\usage{
plot2WR(matrixWR2)
}
\arguments{
\item{matrixWR2}{The two-way reference matrix (2WR) from the Dist2WR function.}
}
\value{
Return a level plot of the 2WR matrix. It can be informative to look at the complete 2WR-matrix before doing any further analysis. It gives a visual idea of the overall congruence or incongruence in the dataset.
}
\description{
This function permits to plot the 2WR matrix.
}
