% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylter.R
\name{print.phylter}
\alias{print.phylter}
\title{Print phylter objects}
\usage{
\method{print}{phylter}(x, ...)
}
\arguments{
\item{x}{Object returned by function 'phylter()'.}

\item{...}{Additional arguments.}
}
\value{
Print formatting
}
\description{
Print on screen a simple description of the content of phylter objects
}
\examples{
data(carnivora)
res<-phylter(carnivora)
print(res)
}
