% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylter.R
\name{print.phylterfinal}
\alias{print.phylterfinal}
\title{print.phylterfinal}
\usage{
\method{print}{phylterfinal}(x, ...)
}
\arguments{
\item{x}{Object returned by function 'phylter()'.}

\item{...}{Additional arguments.}
}
\value{
Print formatting
}
\description{
print.phylterfinal TODO.
}
