% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylter.R
\name{print.phylterinitial}
\alias{print.phylterinitial}
\title{print.phylterinitial}
\usage{
\method{print}{phylterinitial}(x, ...)
}
\arguments{
\item{x}{Object returned by function 'phylter()'.}

\item{...}{Additional arguments.}
}
\value{
Print formatting
}
\description{
print.phylterinitial TODO.
}
