% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylter.R
\name{print.summary.phylter}
\alias{print.summary.phylter}
\title{print summary of phylter objects}
\usage{
\method{print}{summary.phylter}(x, ...)
}
\arguments{
\item{x}{Object returned by function 'summary.phylter()'.}

\item{...}{Additional arguments.}
}
\value{
NA
}
\description{
Prints on screen the summary of an object of class summmary.phylter as returned by the \code{summary.phylter()} function.
}
\examples{
data(carnivora)
res<-phylter(carnivora)
summary<-summary(res)
print(summary)
}
