% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{rm.gene.and.species}
\alias{rm.gene.and.species}
\title{rm.gene.and.species}
\usage{
rm.gene.and.species(trees, sp2rm, gn2rm)
}
\arguments{
\item{trees}{list of gene trees (in multiphylo format) from which we want to remove species or genes}

\item{sp2rm}{species to remove as a list}

\item{gn2rm}{genes to remove as a list}
}
\value{
Return a list of gene trees without the species or genes removed
}
\description{
Suppress species or genes in a list of gene trees.
}
