% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.trees.R
\name{simulate.trees}
\alias{simulate.trees}
\title{simulate.trees}
\usage{
\method{simulate}{trees}(Ngn, Nsp, Nsp.out = 0, Ngn.out = 0,
  Nb.cell.outlier = 0, brlen.sd = 0)
}
\arguments{
\item{Ngn}{Number of gene trees to simulate.}

\item{Nsp}{Number of species (tips) per tree.}

\item{Nsp.out}{Number of outlier species (also called rogue taxa). 0 = none.}

\item{Ngn.out}{Number of outlier genes. 0 = none.}

\item{Nb.cell.outlier}{Number of times one species in one tree is misplaced. 0 = none.}

\item{brlen.sd}{Heterogeneity of branch lengths in trees. A value with mean 0 and standard
deviation equal to brlen.sd is added to each branch length.}
}
\value{
A list of trees in 'multiPhylo' format.
}
\description{
Simple(simplistic) simulation of trees with outliers. No missing data. 
A first tree is generated with rtree() and is then modified according
to the parameters chosen by the user.
}
