% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylter.R
\name{summary.phylter}
\alias{summary.phylter}
\title{summary.phylter}
\usage{
\method{summary}{phylter}(object, ...)
}
\arguments{
\item{object}{Object returned by function 'phylter()'.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Print formatting
}
\description{
summary.phylter TODO.
}
