% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.phylter.R
\name{write.phylter}
\alias{write.phylter}
\title{Write phylter summary to file}
\usage{
write.phylter(x, file = "phylter.out")
}
\arguments{
\item{x}{The object returned by the 'phylter()' function.}

\item{file}{Name of the file where to write the summary of the phylter output.}
}
\description{
This function writes to a file a summary of the phylter analysis.
}
