% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistatisFast.R
\name{DistatisFast}
\alias{DistatisFast}
\title{DistatisFast}
\usage{
DistatisFast(matrices, Norm = TRUE, factorskept = 2)
}
\arguments{
\item{matrices}{A list of K distance matrices, all of the same dimension (IxI).}

\item{Norm}{Should the matrices be normalized. If TRUE (the default), 
each matrix is normalized such that its first eigenvalie is equal to one.}

\item{factorskept}{Number of factors to keep for the computation of the factor 
scores of the observations.}
}
\value{
Returns a list contanining 
\itemize{
	\item 'alpha': array of length K of the weight associated to each matrix.
 \item 'lambda': array of length K of the normalization factors used for each matrix 
 if Norm=TRUE.
	\item 'RVmat': a KxK matrix with RV correlation coefficient computed between all
 pairs of matrices.
	\item 'compromise': an IxI matrix representing the best compromise between all 
matrices. This matrix is the weighted average of all K matrices, using 'alpha' as
a weighting array.
	\item 'quality': the quality of the compromise. This value between 0 and 1 
describes how much of the variance of the K matrices is captured by the compromise.   
}
}
\description{
New implementation of the DISTATIS method for K matrices of dimension IxI.
This version of distatis is faster than the original one because only the minimum required number
of eigenvalues and eigenvectors is calculated. The difference in speed 
is particularly visible when the number of matricesis big.
}
\references{
Abdi, H., Valentin, D., O'Toole, A.J., & Edelman, B. (2005).
DISTATIS: The analysis of multiple distance matrices.
Proceedings of the IEEE Computer Society: International
Conference on Computer Vision and Pattern Recognition_.  (San
Diego, CA, USA). pp. 42-47.
}
