% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(mat, what = "none")
}
\arguments{
\item{mat}{A matrix}

\item{what}{Character string indicating whether the matrix should be normalized and how. If what="none", the matrix is not normalized (the default), if what="species", the matrix is normalized so that the difference between species is increased, and if what="genes", the matrix is normalized so that the difference between genes is increased.}
}
\value{
A normalized matrix
}
\description{
This function normalizes the 2WR matrix (or any matrix) according to the species (rows) or to the genes (columns).
normalize is a function taken from the method phylo-MCOA (de Vienne M.D., Ollier S. et Aguileta G. (2012) Phylo-MCOA: A Fast and Efficient Method to Detect Outlier Genes and Species in Phylogenomics Using Multiple Co-inertia Analysis. Molecular Biology and Evolution 29 : 1587 – 1598)
}
