% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{PhylteR}
\alias{PhylteR}
\title{PhylteR}
\usage{
PhylteR(trees, distance = "patristic", bvalue = 0, method.imp = "IPCA",
ncp = 3, center = FALSE, scale = FALSE, maxiter = 1000, k = 1.5,
thres = 0.5, gene.names = NULL, Norm = "NONE")
}
\arguments{
\item{trees}{The list of gene trees}
\item{distance}{parameter from the function trees2matrices to transform trees into distance matrices. Distance could be "nodal" or "patristic" (default).}
\item{bvalue}{This argument is only used if trees contain bootstrap values. It determines under what bootstrap values the nodes should be collapsed. Value 0 (the default) means that no nodes are collapsed.}
\item{method.imp}{The method used for missing data imputation. "IPCA" for imputation with iteractive PCA (Slower but more accurate) ."MEAN" for imputation by means (faster but less accurate). If there is too much missing data in your data set the iterative PCA ("IPCA") method may not converge. Please then use method.imp ⁼ "MEAN". If there is too much missing data in your data set the iterative PCA ("IPCA") method may not converge. Please then use method.imp ⁼ "MEAN".}
\item{ncp}{only used if method.imp = "IPCA". integer corresponding to the number of components used to to predict the missing entries.}
\item{center}{only used if method.imp = "IPCA". boolean. By default FALSE leading to data not centered.}
\item{scale}{only used if method.imp = "IPCA". boolean. By default FALSE leading to not a same weight for each variable.}
\item{maxiter}{only used if method.imp = "IPCA". integer, maximum number of iteration for the algorithm.}
\item{k}{the strength of outlier assignement. The higher this valu,e the more stringent the detection (less outliers detected).}
\item{thres}{For the detection of complete outlier. Threshold above which genes or species are considered as complete outliers. 0.5 means that a gene or a species is a complete outlier if it is detected as outlier for more than 50\% of the species or genes respectively.}
\item{gene.names}{List of gene names if the user want to renames the list of trees. NULL by default.}
\item{Norm}{Type of normalization used for the function mat2dist. Current options are NONE (default) or MFA (that normalizes each matrix so that its first eigenvalue is equal to one).}
}

\value{
\item{$Complete$mat2WR}{The 2WR matrix used to detect complete outliers.}
\item{$complete$outgn}{The list of complete outliers genes.}
\item{$complete$outsp}{The list of complete outliers species.}
\item{$CellByCell$outcell}{The list of cell outliers.}
}
\description{
This function finds complete and cell outliers inside a list of gene trees.
}
\details{
The detection is done in two steps. The first step is the detection of complete outliers. Complete outliers detected are then removed of the list of trees and the second step is the detection of cell outliers in this list.
}
\examples{

# Detecting outliers of the dataset Fungi using nodal distances.
# This data set doesn't contain any missing data.

data(Fungi)

Results <- PhylteR(Fungi, distance = "nodal", bvalue = 0, k = 3,
thres = 0.6, gene.names = NULL, Norm = "NONE")

# See results
# Complete outliers

outgn <- Results$complete$outgn
outsp <- Results$complete$outsp

# outliers cell

outcell <- Results$CellByCell$outcell

# you can visualize the 2WR matrices (genes x species) with the function plot2WR.

plot = plot2WR(Results$Complete$mat2WR)

}
