% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylter.R
\name{phylter}
\alias{phylter}
\title{phylter}
\usage{
phylter(X, bvalue = 0, distance = "patristic", k = 3, thres = 0.3,
  Norm = TRUE, keep.species = TRUE, gene.names = NULL,
  test.island = TRUE, verbose = TRUE, stop.criteria = 1e-05)
}
\arguments{
\item{X}{A list of phylogenetic trees (phylo object) or a list 
of distance matrices. Trees can have different number of leaves and matrices
can have different dimensions. If this is the case, missing values are imputed.}

\item{bvalue}{If X is a list of trees, nodes with a support below 'bvalue' will be collapsed
prior to the outlier detection.}

\item{distance}{If X is a list of trees, type of distance used to compute the 
pairwise matrices for each tree. Can be "patristic" (sum of branch lengths separating tips, the default)
or nodal (number of nodes separating tips).}

\item{k}{Strength of outlier detection. The higher this value the less outliers
detected (see details).}

\item{thres}{For the detection of complete outliert. Threshold above which genes or species
are considered as complete outliers. 0.3 means that a gene (resp. species) is a
complete outlier if it is detected as outlier for more than 30\% of 
its species (resp. genes).}

\item{Norm}{Should the matrices be normalized. If TRUE (the default), 
each matrix is normalized such that its first eigenvalie is equal to one.}

\item{keep.species}{If TRUE, species are protected from being detected 
as complete outliers and filtered out.}

\item{gene.names}{List of gene names used to rename elements in X. If NULL (the default), 
elements are named 1,2,..,length(X).}

\item{test.island}{This should not be modified. If TRUE (the default), only the highest value in
an 'island' of outliers is considered an outlier. This prevents non-outliers hitchhiked by outliers
to be considered outliers themselves.}

\item{verbose}{If TRUE (the default), messages are written duringt the filtering process to get information
of what is happening}

\item{stop.criteria}{The optimisation stops when the gain between round n and round n+1 is smaller
than this value. Default to 1e-5.}
}
\value{
A list with the 'Initial' (before filtering) and 'Final' (after filtering) states/
}
\description{
Detection and filtering out of outliers in a list of trees 
or distance matrices.
}
