% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylter.R
\name{phylter}
\alias{phylter}
\title{filter phylogenomics datasets}
\usage{
phylter(
  X,
  bvalue = 0,
  distance = "patristic",
  k = 3,
  k2 = k,
  Norm = "median",
  Norm.cutoff = 1e-06,
  gene.names = NULL,
  test.island = TRUE,
  verbose = TRUE,
  stop.criteria = 1e-05,
  InitialOnly = FALSE,
  old = FALSE,
  normalizeby = "row"
)
}
\arguments{
\item{X}{A list of phylogenetic trees (phylo object) or a list 
of distance matrices. Trees can have different number of leaves and matrices
can have different dimensions. If this is the case, missing values are imputed.}

\item{bvalue}{If X is a list of trees, nodes with a support below 'bvalue' will be collapsed
prior to the outlier detection.}

\item{distance}{If X is a list of trees, type of distance used to compute the 
pairwise matrices for each tree. Can be "patristic" (sum of branch lengths separating tips, the default)
or nodal (number of nodes separating tips).}

\item{k}{Strength of outlier detection. The higher this value the less outliers
detected (see details).}

\item{k2}{Same as k for complete gene outlier detection. To preserve complete genes from 
being discarded, k2 can be increased . By default, k2 = k. (see above) 
By default, k2=k.}

\item{Norm}{Should the matrices be normalized and how. If "median", matrices are divided by their median, if 
"mean" they are divided by their mean, if "none", no normalization if performed. Normalizing ensures that fast-evolving 
(and slow-evolving) genes are not treated as outliers. Normalization by median is less sensitive to outlier values
but can lead to errors if some matrices have a median value of 0. 
are not considered outliers.}

\item{Norm.cutoff}{Value of the median (if Norm="median") or the mean (if
Norm="mean") below which matrices are simply discarded from the analysis. This
prevents dividing by 0, and getting rid of genes that contain mostly branches
of length 0 and are therefore uninformative anyway.}

\item{gene.names}{List of gene names used to rename elements in X. If NULL (the default), 
elements are named 1,2,..,length(X).}

\item{test.island}{This should not be modified. If TRUE (the default), only the highest value in
an 'island' of outliers is considered an outlier. This prevents non-outliers hitchhiked by outliers
to be considered outliers themselves.}

\item{verbose}{If TRUE (the default), messages are written during the filtering process to get information
of what is happening}

\item{stop.criteria}{The optimisation stops when the gain between round n and round n+1 is smaller
than this value. Default to 1e-5.}

\item{InitialOnly}{Logical. If TRUE, only the Initial state of teh data is computed. The optimization and}

\item{old}{Logical. Should the old way of detecting outliers be used. Default to FALSE.
outlier detection is NOT performed. Useful to get an idea about the initial state of th data.}

\item{normalizeby}{Should the 2WR matrix be normalized prior to outlier detection, and how.}
}
\value{
A list of class 'phylter' with the 'Initial' (before filtering) and 'Final' (after filtering) states, 
or a list of class 'phylterinitial' only, if InitialOnly=TRUE.
}
\description{
Detection and filtering out of outliers in a list of trees 
or distance matrices.
}
