% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.phylter.R
\name{plot.phylter}
\alias{plot.phylter}
\title{plot.phylter}
\usage{
\method{plot}{phylter}(X, what = "all", layout = 1)
}
\arguments{
\item{X}{The object returned by the 'phylter' function.}

\item{show.missing}{Logical. Should missing data be represented by white dots?}

\item{show.outliers}{Logical. Should outliers be represented by black dots on the final matrix?}

\item{transpose}{Logical. If FALSE (the default), species are in rows and genes in columns. If TRUE, 
rows and columns are inverted.}
}
\value{
A plot of the 2WR matrix.
}
\description{
Plot the initial and final 2WR matrices side by side. Highlights missing data and detected outliers
}
