% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{plot2WR}
\alias{plot2WR}
\title{plot2WR}
\usage{
plot2WR(matrixWR2)
}
\arguments{
\item{matrixWR2}{The two-way reference matrix (2WR) from the Dist2WR function.}
}
\value{
Return a level plot of the 2WR matrix. It can be informative to look at the complete 2WR-matrix before doing any further analysis. It gives a visual idea of the overall congruence or incongruence in the dataset.
}
\description{
This function permits to plot the 2WR matrix.
}
\examples{
# Detecting outliers of the dataset Fungi using nodal distances.
# This data set doesn't contain any missing data.

data(Fungi)

Results <- PhylteR(Fungi, distance = "nodal", bvalue = 0, k = 3,
thres = 0.6, gene.names = NULL, Norm = "NONE")

# you can visualize the 2WR matrices (genes x species) with the function plot2WR.

plot = plot2WR(Results$Complete$mat2WR)
}
