% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylter.R
\name{print.phylter}
\alias{print.phylter}
\title{print.phylter}
\usage{
\method{print}{phylter}(x, ...)
}
\arguments{
\item{x}{Object returned by function 'phylter()'.}

\item{...}{Additional arguments.}
}
\value{
Print formatting
}
\description{
print.phylter TODO.
}
