% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylter.R
\name{print.phylterfinal}
\alias{print.phylterfinal}
\title{print objects of class phylterfinal}
\usage{
\method{print}{phylterfinal}(x, ...)
}
\arguments{
\item{x}{Object returned by function 'phylter()'.}

\item{...}{Additional arguments.}
}
\value{
NA
}
\description{
Print on screen a simple description of the content of objects of class \code{phylterfinal}.
}
\examples{
data(carnivora)
res<-phylter(carnivora)
print(res$Final)
}
