% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylter.R
\name{print.summary.phylter}
\alias{print.summary.phylter}
\title{print.summary.phylter}
\usage{
\method{print}{summary.phylter}(x)
}
\arguments{
\item{X}{Object returned by function 'phylter()'.}
}
\value{
Print formatting
}
\description{
print.summary.phylter TODO.
}
