% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{rename.genes}
\alias{rename.genes}
\title{rename.genes}
\usage{
rename.genes(trees, gene.names = NULL)
}
\arguments{
\item{trees}{A list of gene trees in multiphylo format}

\item{gene.names}{List of genes names the user wants to give to the list of trees. It should be of the same lenght of the list of trees. If NULL, genes are numeroted from 1 to the number of genes.}
}
\value{
The list of renamed trees in multiphylo format.
}
\description{
This function permits the user to add names to the genes trees.
}
