% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylter.R
\name{summary.phylter}
\alias{summary.phylter}
\title{summary.phylter}
\usage{
\method{summary}{phylter}(X)
}
\arguments{
\item{X}{Object returned by function 'phylter()'.}
}
\value{
Print formatting
}
\description{
summary.phylter TODO.
}
