% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylter.R
\name{summary.phylter}
\alias{summary.phylter}
\title{Get summary for phylter objects}
\usage{
\method{summary}{phylter}(object, ...)
}
\arguments{
\item{object}{Object returned by function 'phylter()'.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
On object of class \code{summmary.phylter}
}
\description{
Get the summary of an object of class \code{phylter}.
}
\examples{
data(carnivora)
res<-phylter(carnivora)
summary<-summary(res)
class(summary)
}
