% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylteR.R
\name{trees2matrices}
\alias{trees2matrices}
\title{trees2matrices}
\usage{
trees2matrices(trees, distance = "patristic", bvalue = 0)
}
\arguments{
\item{trees}{A list of gene trees in multiphylo format.}

\item{distance}{A method to generate distance matrices. It could be "nodal" to establish that the distance between two species is the number of nodes that separate them. Or "patristic" (default) if the distance between two species is be the sum of branch lengths between them.}

\item{bvalue}{This argument is only used if trees contain bootstrap values. It determines under what bootstrap values the nodes should be collapsed. Value 0 (the default) means that no nodes are collapsed.}
}

\value{
return a list of distance matrices
}
\description{
trees2matrices changes a list of trees into a list of matrices.
}
